Restructuring + layout (#19)
