+++ ui/src/components/Navigator/index.ts
@@ -1,0 +1,1 @@
+export { Navigator } from './Navigator';
