+++ ui/src/components/index.ts
@@ -1,4 +1,6 @@
-export { Header } from '../layouts/Header';
 export { Loading } from './Loading';
 export { OAuthCallback } from './OAuthCallback';
 export { UserMenu } from './UserMenu';
+export { Login } from './Login';
+export { Header } from './Header';
+export { Navigator } from './Navigator';
