--- ui/src/layouts/NavBar/NavBar.css
@@ -1,95 +1,0 @@
-.navbar {
-    background-color: var(--mantine-color-body);
-    height: 700px;
-    padding: var(--mantine-spacing-md);
-    padding-top: 0;
-    display: flex;
-    flex-direction: column;
-    border-right: 1px solid var(--mantine-color-gray-3);
-}
-
-.section {
-    margin-left: calc(var(--mantine-spacing-md) * -1);
-    margin-right: calc(var(--mantine-spacing-md) * -1);
-    margin-bottom: var(--mantine-spacing-md);
-
-    &:not(:last-of-type) {
-        border-bottom: 1px solid
-       var(--mantine-color-gray-3);
-    }
-}
-
-.searchCode {
-    font-weight: 700;
-    font-size: 10px;
-    background-color: var(--mantine-color-gray-0);
-    border: 1px solid var(--mantine-color-gray-2);
-}
-
-.mainLinks {
-    padding-left: calc(var(--mantine-spacing-md) - var(--mantine-spacing-xs));
-    padding-right: calc(var(--mantine-spacing-md) - var(--mantine-spacing-xs));
-    padding-bottom: var(--mantine-spacing-md);
-}
-
-.mainLink {
-    display: flex;
-    align-items: center;
-    width: 100%;
-    font-size: var(--mantine-font-size-xs);
-    padding: 8px var(--mantine-spacing-xs);
-    border-radius: var(--mantine-radius-sm);
-    font-weight: 500;
-    color: var(--mantine-color-gray-7);
-
-    &:hover {
-        background-color: var(--mantine-color-gray-0);
-        color: var(--mantine-color-gray-7);
-    }
-}
-
-.mainLinkInner {
-    display: flex;
-    align-items: center;
-    flex: 1;
-}
-
-.mainLinkIcon {
-    margin-right: var(--mantine-spacing-sm);
-    color: var(--mantine-color-gray-6);
-}
-
-.mainLinkBadge {
-    padding: 0;
-    width: 20px;
-    height: 20px;
-    pointer-events: none;
-}
-
-.collections {
-    padding-left: calc(var(--mantine-spacing-md) - 6px);
-    padding-right: calc(var(--mantine-spacing-md) - 6px);
-    padding-bottom: var(--mantine-spacing-md);
-}
-
-.collectionsHeader {
-    padding-left: calc(var(--mantine-spacing-md) + 2px);
-    padding-right: var(--mantine-spacing-md);
-    margin-bottom: 5px;
-}
-
-.collectionLink {
-    display: block;
-    padding: 8px var(--mantine-spacing-xs);
-    text-decoration: none;
-    border-radius: var(--mantine-radius-sm);
-    font-size: var(--mantine-font-size-xs);
-    color: var(--mantine-color-gray-7);
-    line-height: 1;
-    font-weight: 500;
-
-    &:hover {
-        background-color: var(--mantine-color-gray-0);
-        color: var(--mantine-color-gray-7);
-    }
-}
