--- ui/src/layouts/NavBar/NavBar.tsx
@@ -1,3 +1,0 @@
-import './NavBar.css';
-
-export const NavBar = () => <nav></nav>;
