--- ui/src/layouts/NavBar/index.ts
@@ -1,1 +1,0 @@
-export { NavBar } from './NavBar';
